Add files via upload
This method of learning maintains the attendance of buses and class rooms through scanning the barcode that is present on the students identity cards. This scans the barcode number and lists it in a separate CSV file. Then this CSV file data is checked against the strength of the bus or class. Also, it provides the attendance list as one CSV file.
Here, the barcode scanner is a mobile camera, which is connected through the IP address of the mobile.